feat(suite-native): extract accordion content for reuse
